+++ src/main/java/pl/jdacewicz/socialmediaserver/reaction/dto/ReactionCounter.java
@@ -1,5 +1,5 @@
 package pl.jdacewicz.socialmediaserver.reaction.dto;
 
 public record ReactionCounter(ReactionDto reaction,
-                              int count) {
+                              long count) {
 }
